+++ apps/backend/src/libs/enums/enums.ts
@@ -1,8 +1,11 @@
+export { DefaultMaxFileSize } from "./default-max-file-size.enum.js";
 export { FastifyHook } from "./fastify-hook.enum.js";
 export {
 	APIPath,
 	AppEnvironment,
 	AuthApiPath,
 	ExceptionMessage,
+	S3ErrorMessage,
 	ServerErrorType,
+	UploadErrorMessage,
 } from "@meetlytic/shared";