+++ apps/backend/src/modules/users/libs/types/delete-avatar-options.type.ts
@@ -1,0 +1,9 @@
+import { type APIHandlerOptions } from "~/libs/modules/controller/controller.js";
+
+import { type UserResponseDto } from "../../libs/types/types.js";
+
+type DeleteAvatarOptions = APIHandlerOptions<{
+	user: UserResponseDto;
+}>;
+
+export { type DeleteAvatarOptions };